migration aggregators to extra package Signed-off-by: dbulahov <bulahovdenis@gmail.com>
